Fix for XMLRPC issue.  We no longer depend on an XML parse.  It was unsuccessful sometimes.  Instead the code does a quick regex parse of the returned web page text.
This commit also includes a modalViewController "pop-up" for cases where we can't hit the xmlrpc endpoint.  The view gives the user the opportunity to manually enter their xmlrpc endpoint url.

This was a bit tricky to test - I left one line of code (commented out) in BlogDataManager:refreshCurrentBlog that serves as a test by setting the xmlrpc value to nil so that the view that requests the manual entry will get shown.  Future testers will want to look for this line... at the time of this writing, line number 1110 in BlogDataManager

	//Unique search string: ^^XMLRPC Endpoint
	//xmlrpc = nil;