Add support for self-signed certificates
Instead of just throwing a generic error, we give the option to continue connecting.
I think the current API for iOS doesn't allow us to store trust permanently, so the alert will show again when the app relaunches.

It's not ideal but at least sites with self-signed or invalid certificates can be used if the user wants to.

fixes #1306