Photo/Settings button updates and fixes.
* Better font resizing for languages with long strings
* New button style with a bigger tap area, fixed icons and margins
* Updated uploading animations/styling
* Updated resources overall

Tested on devices iphone 5 + ipad both on iOS 6.

Also there's a fix in here for hiding the sidebar etc when the Welcomeviewcontroller is showing, but it crashes every time you try to add a WPcom blog to the app - however with the fix removed it still does it so needs further investigation.